# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RiceBook API",
    "version": "1.0.0",
    "description": "RiceBook 后端 API 文档"
  },
  "paths": {
...
    "/headline/{user}": {
      "get": {
        "summary": "Get the headline for a user",
        "description": "Retrieve the headline of a user by username. If no username is provided, get the logged-in user's headline.",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Username of the user"
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Headline retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "headline": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/headline": {
      "put": {
        "summary": "Update the logged-in user's headline",
        "description": "Update the headline for the logged-in user.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "headline": {
                    "type": "string"
                  }
                },
                "required": [
                  "headline"
                ]
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Headline updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "headline": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Headline is required"
          },
          "404": {
            "description": "Profile not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/email/{user}": {
      "get": {
        "summary": "Get the email for a user",
        "description": "Retrieve the email of a user by username. If no username is provided, get the logged-in user's email.",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Username of the user"
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Email retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/email": {
      "put": {
        "summary": "Update the logged-in user's email",
        "description": "Update the email for the logged-in user.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  }
                },
                "required": [
                  "email"
                ]
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Email updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Email is required"
          },
          "404": {
            "description": "Profile not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/password": {
      "put": {
        "summary": "Change the logged-in user's password",
        "description": "Update the password for the logged-in user. The new password will be hashed and stored securely.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "password"
                ]
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Password updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "New password is required"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
...
}